[bug] Support anything that inherits File
This was creating a problem as we needed the metadata from Bamboo to be able to properly name the file when we uploaded it to ChartHop. In doing that, I made a new class that inherited from File, but that would not get properly added to FormData because the comparison is class specific vs type specific. The `is` check will allow anything that inherits from File.